Implementing decorate command
