+++ cmd/decorate_test.go
@@ -1,0 +1,10 @@
+package cmd
+
+import "testing"
+
+func Test_NewDecorateCmd_CheckReturn(t *testing.T) {
+	cmd := NewDecorateCmd()
+	if cmd == nil {
+		t.Errorf("NewDecorateCmd() = nil")
+	}
+}
